fix(data): fix french indexing
the french field in title use a custom anaylyser that add a marker at
the beginning of the string. This allow us to have a match_prase_prefix
that match the start of the string.

We use french_indexing as analyser to add our token at index time and
use the normal french analyser at search time (search_analyser)
since we add our token at the api level.